Backend/requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
python-dotenv==1.0.1
waitress==3.0.0